requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-fsm==2.7.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Pillow==8.4.0
PyJWT==2.3.0
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
